Update: Refactor tiny mce
